requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
altgraph=0.17=py_0
appdirs=1.4.4=py_0
attrs=20.1.0=py_0
black=19.10b0=py_0
blas=1.0=mkl
ca-certificates=2020.7.22=0
certifi=2020.6.20=py37_0
click=7.1.2=py_0
dlib=19.21.0=py37he05b8f6_0
freetype=2.10.2=hd328e21_0
future=0.18.2=py37_1
hdf5=1.8.20=hac2f561_1
icc_rt=2019.0.0=h0cc432a_1
imutils=0.5.3=pypi_0
intel-openmp=2020.2=254
jpeg=9d=he774522_0
libopencv=3.4.2=h20b85fd_0
libpng=1.6.37=h2a8f88b_0
libtiff=4.1.0=h56a325e_1
lz4-c=1.9.2=h62dcd97_1
macholib=1.14=py_1
mkl=2020.2=256
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.1.0=py37h45dec08_0
mkl_random=1.1.1=py37h47e9c7a_0
mypy_extensions=0.4.3=py37_0
numpy=1.19.1=py37h5510c5b_0
numpy-base=1.19.1=py37ha3acd2a_0
olefile=0.46=py37_0
opencv=3.4.2=py37h40b0b35_0
openssl=1.1.1g=he774522_1
pathspec=0.7.0=py_0
pefile=2019.4.18=py_0
pillow=7.2.0=py37hcc1f983_0
pip=20.2.2=py37_0
py-opencv=3.4.2=py37hc319ecb_0
pycryptodome=3.8.2=py37he774522_0
pyinstaller=3.6=py37h62dcd97_5
pyqt5=5.15.0=pypi_0
pyqt5-sip=12.8.0=pypi_0
python=3.7.7=h81c818b_4
python_abi=3.7=1_cp37m
pywin32=227=py37he774522_1
pywin32-ctypes=0.2.0=py37_1001
regex=2020.7.14=py37he774522_0
scipy=1.5.2=py37h9439919_0
setuptools=49.6.0=py37_0
six=1.15.0=py_0
sqlite=3.33.0=h2a8f88b_0
tk=8.6.10=he774522_0
toml=0.10.1=py_0
typed-ast=1.4.1=py37he774522_0
typing_extensions=3.7.4.3=py_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_3
wheel=0.35.1=py_0
wincertstore=0.2=py37_0
xz=5.2.5=h62dcd97_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.5=h04227a9_0
